cleanup not needed everytime -->
--> just on one build of the matrix

+++ tools/travis_pushAndCloseRepo.sh
@@ -2,12 +2,16 @@
 cd rte-www
 echo "Copying logs"
 cp ../rte/logs/* ./logs/.
-echo "Cleaning up log dir (leave the 50 newest)"
-mkdir ./logs/tmpSafe
-cp -v `find ./logs/ -maxdepth 1  -name 'acplt_build*.log' -type f -printf '%T@ %p\n' | sort -n | tail -50 | cut -f2- -d" " ` ./logs/tmpSafe/.
-rm -v `find ./logs/ -maxdepth 1 -name 'acplt_build*.log' -type f`
-mv -v ./logs/tmpSafe/* ./logs/.
-rm -vR ./logs/tmpSafe
+
+if [[ ${OV_ARCH_BITWIDTH} == "32" && ${CC} == "gcc" ]]; then 
+  echo "Cleaning up log dir (leave the 50 newest)"
+  mkdir ./logs/tmpSafe
+  cp -v `find ./logs/ -maxdepth 1  -name 'acplt_build*.log' -type f -printf '%T@ %p\n' | sort -n | tail -50 | cut -f2- -d" " ` ./logs/tmpSafe/.
+  rm -v `find ./logs/ -maxdepth 1 -name 'acplt_build*.log' -type f`
+  mv -v ./logs/tmpSafe/* ./logs/.
+  rm -vR ./logs/tmpSafe
+fi
+
 git add -A ./logs/*
 echo "Pushing to github"
 git config --global user.email "rte_bot@outlook.de"
